+++ receiver/jaegerreceiver/trace_receiver.go
@@ -24,6 +24,7 @@ import (
 	"net"
 	"net/http"
 	"sync"
+	"time"
 
 	apacheThrift "github.com/apache/thrift/lib/go/thrift"
 	"github.com/gorilla/mux"
@@ -65,13 +66,14 @@ type configuration struct {
 	CollectorGRPCPort           int
 	CollectorGRPCServerSettings configgrpc.GRPCServerSettings
 
-	AgentCompactThriftPort       int
-	AgentCompactThriftConfig     ServerConfigUDP
-	AgentBinaryThriftPort        int
-	AgentBinaryThriftConfig      ServerConfigUDP
-	AgentHTTPPort                int
-	RemoteSamplingClientSettings configgrpc.GRPCClientSettings
-	RemoteSamplingStrategyFile   string
+	AgentCompactThriftPort                   int
+	AgentCompactThriftConfig                 ServerConfigUDP
+	AgentBinaryThriftPort                    int
+	AgentBinaryThriftConfig                  ServerConfigUDP
+	AgentHTTPPort                            int
+	RemoteSamplingClientSettings             configgrpc.GRPCClientSettings
+	RemoteSamplingStrategyFile               string
+	RemoteSamplingStrategyFileReloadInterval time.Duration
 }
 
 // Receiver type is used to receive spans that were originally intended to be sent to Jaeger.
@@ -484,6 +486,7 @@ func (jr *jReceiver) startCollector(host component.Host) error {
 		// init and register sampling strategy store
 		ss, gerr := staticStrategyStore.NewStrategyStore(staticStrategyStore.Options{
 			StrategiesFile: jr.config.RemoteSamplingStrategyFile,
+			ReloadInterval: jr.config.RemoteSamplingStrategyFileReloadInterval,
 		}, jr.settings.Logger)
 		if gerr != nil {
 			return fmt.Errorf("failed to create collector strategy store: %v", gerr)