+++ CHANGELOG.md
@@ -2,14 +2,23 @@
 
 ## Unreleased
 
+## 💡 Enhancements 💡
+
+- `prometheusremotewriteexporter`: Handling Staleness flag from OTLP (#6679)
+
 ## 🛑 Breaking changes 🛑
 
 - `memcachedreceiver`: Update metric names (#6594)
+- `sapm` receiver: Use Jaeger status values instead of OpenCensus (#6682)
+- `jaeger` receiver/exporter: Parse/set Jaeger status with OTel spec values (#6682)
+
 ## 🚀 New components 🚀
 
 ## 🧰 Bug fixes 🧰
 
 ## 💡 Enhancements 💡
+
+- `lokiexporter`: add complete log record to body (#6619)
 
 ## v0.41.0